# backend/api/openapi/api-v1.yaml
openapi: 3.0.0
info:
  title: Open Source ERP API
  description: Api schema for ERP backend.
  version: 0.0.1

servers:
  - url: https://localhost:8080/api/v1
    description: dev url
paths:
  /users:
    get:
      summary: Get a list of users
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Page number of the users list
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of users per page
      responses:
        "200":
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: New user created

  /users/{userId}:
    get:
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A single user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      summary: Update a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: User updated
    delete:
      summary: Delete a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted

components:
  schemas:
    User:
      type: object
      required:
        - username
        - email
        - password
        - status
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        status:
          type: string
          enum: [active, inactive]
        role:
          type: string
          enum: [admin, user, manager]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time